Move initial GET endpoints into an app file for TDD
